back to original graph
