+++ clean_code/clean_code.pyproj
@@ -49,9 +49,6 @@
     <Compile Include="position.py">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="test_missing_number.py">
-      <SubType>Code</SubType>
-    </Compile>
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
   <!-- Uncomment the CoreCompile target to enable the Build command in
